API Reference

API Reference

Module

Basic Types

Specific YCbCr Types (Aliases)

Functions

Module

YUVColorTypes — Module

YUVColorTypes

An add-on libray to ColorTypes supporting YUV-like colors.

Type Hierarchy

ColorTypes.Color{T,N}

AbstractYUV{T}

YUV{T,P,U,R,D}

AbstractLuma{T}

Luma{T,P,U,R,D}

ColorTypes.AlphaColor{C,T,N}

AbstractAYUV{C,T}

AYUV{T,P,U,R,D}

AbstractALuma{C,T}

ALuma{T,P,U,R,D}

ColorTypes.ColorAlpha{C,T,N}

AbstractYUVA{C,T}

YUVA{T,P,U,R,D}

AbstractLumaA{C,T}

LumaA{T,P,U,R,D}

Supported Profiles

YCbCrBT601_625

YCbCrBT601_525

Basic Types

YUVColorTypes.AbstractYUV — Type

AbstractYUV{T} <: Color{T,3}

An abstract type for opaque YUV-like colors.

YUVColorTypes.TransparentYUV — Type

TransparentYUV{C<:AbstractYUV,T} = TransparentColor{C,T,4}

An alias of abstract type TransparentColor for transparent YUV-like colors.

Missing docstring.

Missing docstring for AbstractAYUV. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractYUVA. Check Documenter's build log for details.

Missing docstring.

Missing docstring for `AbstractYUVColorant

YUVColorTypes.YUV — Type

YUV{T,P,U,R,D} <: AbsrtractYUV{T}

T: component type

P: symbol of profile

U: unsigned chroma components if true

R: full range if true

D: bit-depth

See also AYUV and YUVA for transparent versions, and Luma for the single Y component version.

YUVColorTypes.AYUV — Type

AYUV{T,P,U,R,D} <: AbstractAYUV{YUV{T,P,U,R,D},T}

YUVColorTypes.YUVA — Type

YUVA{T,P,U,R,D} <: AbstractYUVA{YUV{T,P,U,R,D},T}

YUVColorTypes.Luma — Type

Luma{T,P,U,R,D} <: AbstractLuma{T}

The single Y component, i.e. grayscale version of a YUV-like color.

YUVColorTypes.ALuma — Type

ALuma{T,P,U,R,D} <: AbstractALuma{Luma{T,P,U,R,D},T}

YUVColorTypes.LumaA — Type

LumaA{T,P,U,R,DR,} <: AbstractLumaA{YUV{T,P,U,R,D},T}

Specific YCbCr Types (Aliases)

YUVColorTypes.YCbCrBT601_625 — Type

YCbCrBT601_625{T,D}

The 625-line version of BT.601 YCbCr color type (previously designated PAL).

YUVColorTypes.YCbCr30BT601_625 — Type

YCbCr30BT601_625{T}

10-bit representation of the 625-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 1024).

YUVColorTypes.YCbCr24BT601_625 — Type

YCbCr24BT601_625{T}

8-bit representation of the 625-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 256). See also YCbCr30BT601_625.

YUVColorTypes.YCbCr36BT601_625 — Type

YCbCr36BT601_625{T}

12-bit representation of the 625-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 4096). See also YCbCr30BT601_625.

YUVColorTypes.YCbCr48BT601_625 — Type

YCbCr48BT601_625{T}

16-bit representation of the 625-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 65536). See also YCbCr30BT601_625.

YUVColorTypes.YCbCrBT601_525 — Type

YCbCrBT601_525{T,D}

The 525-line version of BT.601 YCbCr color type (previously designated NTSC).

YUVColorTypes.YCbCr30BT601_525 — Type

YCbCr30BT601_525{T}

10-bit representation of the 525-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 1024).

YUVColorTypes.YCbCr24BT601_525 — Type

YCbCr24BT601_525{T}

8-bit representation of the 525-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 256). See also YCbCr30BT601_525.

YUVColorTypes.YCbCr36BT601_525 — Type

YCbCr36BT601_525{T}

12-bit representation of the 525-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 4096). See also YCbCr30BT601_525.

YUVColorTypes.YCbCr48BT601_525 — Type

YCbCr48BT601_525{T}

16-bit representation of the 525-line BT.601 YCbCr color type. Its raw components of type T usually have a value in [0, 65536). See also YCbCr30BT601_525.

Functions

YUVColorTypes.luma — Function

luma(c)

Return the luma component of a YUV-like color. To avoid confusion, this function does not support ColorTypes.YCbCr and its variants, which are not under the AbstractYUV.

YUVColorTypes.chroma_u — Function

chroma_u(c)

Return the first chroma component of a YUV-like color.

YUVColorTypes.chroma_v — Function

chroma_v(c)

Return the second chroma component of a YUV-like color.

YUVColorTypes.chroma_b — Function

chroma_b(c)

Return the Cb component of a YCbCr color. For general YUV-like colors, use chroma_u. To avoid confusion, this function does not support ColorTypes.YCbCr and its variants, which are not under the AbstractYUV.

YUVColorTypes.chroma_r — Function

chroma_r(c)

Return the Cr component of a YCbCr color. For general YUV-like colors, use chroma_v. To avoid confusion, this function does not support ColorTypes.YCbCr and its variants, which are not under the AbstractYUV.

YUVColorTypes.chroma_g — Function

chroma_g(c)

Return the Cg component of a YCgCo color. For general YUV-like colors, use chroma_u.

YUVColorTypes.chroma_o — Function

chroma_o(c)

Return the Co component of an YCgCo color. For general YUV-like colors, use chroma_v.

YUVColorTypes.rgb_to_yuv — Function

rgb_to_yuv(C, rgb::AbstractRGB) -> C
rgb_to_yuv(C, argb::TransparentRGB) -> C

Convert an RGB color to a YUV color. The color primaries, whitepoint and gamma are depends on the YUV color type. If you need sRGB colors, use yuv_to_srgb or simply use convert or RGB constructors.

See also srgb_to_yuv and yuv_to_rgb.

YUVColorTypes.srgb_to_yuv — Function

srgb_to_yuv(C, rgb::AbstractRGB) -> C
srgb_to_yuv(C, argb::TransparentRGB) -> C

Convert an sRGB color to a YUV color.

See also rgb_to_yuv and yuv_to_srgb.

YUVColorTypes.yuv_to_rgb — Function

yuv_to_rgb(C, yuv::AbstractYUVColorant) -> C

Convert a YUV color to an RGB color. The color primaries, whitepoint and gamma are depends on the YUV color type. If you need sRGB colors, use yuv_to_srgb or simply use convert or RGB constructors.

See also yuv_to_srgb and rgb_to_yuv .

YUVColorTypes.yuv_to_srgb — Function

yuv_to_srgb(C, yuv::AbstractYUVColorant) -> C

Convert a YUV color to an sRGB color.

See also yuv_to_rgb and srgb_to_yuv.